style: Login, SignUp Modify
