+++ src/components/signup/sizeinput/infobox/InfoBox.tsx
